Updating storage tables to reflect latest code

+++ data/storage/mysql/create_mediator_tables.sql
@@ -6,7 +6,7 @@ CREATE TABLE `rated_cdrs` (
   `id` int(11) NOT NULL AUTO_INCREMENT,
   `cgrid` char(40) NOT NULL,
   `subject` varchar(64) NOT NULL,
-  `cost` double(20,4) DEFAULT NULL,
+  `cost` DECIMAL(20,4) DEFAULT NULL,
   `extra_info` text,
   PRIMARY KEY (`id`),
   UNIQUE KEY `costid` (`cgrid`,`subject`)
